Valid number to 2 decimal places
Couldn't solve this on my one so I'm just posting the solution i was closest too. This was posted on CodeWars by anter69 and it has a good explanation of what everything in the code does even the regex. Her is a link to the python documentation on regex: https://docs.python.org/3/howto/regex.html . This is more for personal use than anything else.